pyproject.toml
[tool.poetry]
name = "django-finance-tracker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = ">=4.2,<5.0"
django-crispy-forms = "^2.3"
django-allauth = "^65.0.2"
django-filter = "^24.3"
django-htmx = "^1.19.0"
crispy-bootstrap5 = "^2024.2"
django-environ = "^0.12.0"
django-import-export = {extras = ["xlsx"], version = "^4.3.7"}


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.4.6"
django-extensions = "^3.2.3"
ruff = "^0.6.8"
ipython = "^8.28.0"
faker = "^37.0.2"


[tool.poetry.group.test.dependencies]
coverage = "^7.8.0"
factory-boy = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
extend-exclude = ["migrations"]

[tool.ruff.lint]
extend-select = ["I"]
